% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arma_mvrnorm}
\alias{arma_mvrnorm}
\title{Generating a multivariate gaussian distribution using RcppArmadillo}
\usage{
arma_mvrnorm(n, mu, sigma)
}
\arguments{
\item{n}{- the number of samples}

\item{mu}{- the mean vector}

\item{sigma}{- the covariace matrix}
}
\description{
There are many ways to simulate a multivariate gaussian distribution
assuming that you can simulate from independent univariate normal
distributions. One of the most popular method is based on the Cholesky
decomposition.
}
\note{
Author: Ahmadou Dicko — written Mar 12, 2013
License: GPL

Reference:
http://gallery.rcpp.org/articles/simulate-multivariate-normal/
}
