% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arma_rgig}
\alias{arma_rgig}
\title{Draw sample from the Generalized Inverse Gaussian (GIG) distribution.}
\usage{
arma_rgig(n, lambda, chi, psi)
}
\arguments{
\item{n}{sample size (positive integer)}

\item{lambda}{parameter for distribution}

\item{chi}{parameter for distribution}

\item{psi}{parameter for distribution}
}
\value{
random sample of size 'n'
}
\description{
Draw sample from the Generalized Inverse Gaussian (GIG) distribution.
}
\note{
Adapted from the R package GIGrvg.

All rights reserved with the R packge authors:
Josef Leydold and Wolfgang Hörmann.
}
\seealso{
Other utils: \code{\link{sample_antoniak}},
  \code{\link{sample_dirichlet}},
  \code{\link{sample_multinomial}}
}
