% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spmm_ags_hmc}
\alias{spmm_ags_hmc}
\title{Hamiltonian Monte Carlo within Augmented Gibbs sampler for Gaussian
Shrinkage Partial Membership models}
\usage{
spmm_ags_hmc(K, alpha, lambda, sigma, max_iter, epsilon, Y, verbose)
}
\arguments{
\item{K}{Number of components or subtypes}

\item{alpha}{Hyperparameter for the Dirichlet prior}

\item{lambda}{Variance smoothing parameter for the normal prior}

\item{sigma}{Variance parameter for the data generating normal}

\item{max_iter}{Maximum number of iterations for the HMC within AGS
algorithm}

\item{epsilon}{The step size of the HMC leep-frog iteration}

\item{Y}{The data matrix, an \code{n_samples x M} matrix, where
\code{n_samples} is the number of observations and \code{M} is the
number of features or genes}

\item{verbose}{0 - no output; 1 - some output; 2 - more output}
}
\description{
A GSPM model is indexed by scalar parameters K, alpha, lambda, and sigma
}
\note{
Author: Clint P. George

License: GPL-3
}
\seealso{
Other mcmc: \code{\link{gspm_ags_hmc}}
}
